Move String.mastodonLength to StringUtils
Removes the need for a synthetic accessor
